P2-01: Mutation and genetic distance
Goal: implement SPEC §4.6 mutation (small, rare-large, slow hue) and
trait-block Euclidean distance, and make every birth mutate.

Tests: test/unit/genome.test.js (+11 cases: mutation bounds, sd checks
for small/large/hue-scaled mutation, a no-op case, determinism, distance
symmetry/zero/weight-exclusion, MAX_TRAIT_DISTANCE), test/unit/breeding.test.js
(renamed the exact-copy case to disable mutation explicitly, added a
differs-by-default case). All new cases passed on the first attempt.

Config keys introduced: genome.sigmaMut (0.05), genome.pMut (0.15),
genome.pBig (0.01), genome.hueScale (0.2), all ⚠️ assumptions.

Also fixed reflex.js's act(): the bounds check compared unrounded doubles
against width/height, then stored those same doubles into store.x/y
(Float32Array, which rounds on assignment) -- a double just under the
boundary could round up to exactly it once stored. Mutation's added rng
draws per birth reshuffled seed 3's downstream sequence enough to finally
hit this always-latent P1-06 edge case in bounds.test.js. Fixed by
rounding with Math.fround before the check.

Phone: n/a.

+++ src/core/reflex.js
@@ -120,8 +120,14 @@ export function act(world, i) {
   const speed = store.pheno[i * TRAIT_COUNT + TRAIT.speed];
   const v = (speed * throttle) / moveCost;
 
-  const nx = store.x[i] + cos(heading) * v;
-  const ny = store.y[i] + sin(heading) * v;
+  // Rounded to float32 *before* the bounds check, not after: store.x/y are
+  // Float32Array, so an unrounded double just under width/height (passing
+  // the check) can round up to exactly width/height once stored, putting
+  // the organism out of bounds (found via P2-01's bounds.test.js — adding
+  // mutation's rng draws shifted seed 3's whole sequence enough to finally
+  // hit this always-latent P1-06 edge case).
+  const nx = Math.fround(store.x[i] + cos(heading) * v);
+  const ny = Math.fround(store.y[i] + sin(heading) * v);
 
   if (nx < 0 || nx >= world.width || ny < 0 || ny >= world.height) {
     store.heading[i] = wrapAngle(heading + TAU / 2);
